refactor supervisorReport
